* Moved BucketUtils to BucketModels, as it's a more descriptive name. * Moved model handling code from OnnxLocalService into new LocalModels, to keep pattern.
