+++ synt/synt-dagpenger/src/main/java/no/nav/dolly/synt/dagpenger/onnx/OnnxBucketService.java
@@ -4,7 +4,7 @@ import jakarta.annotation.PostConstruct;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import no.nav.dolly.synt.dagpenger.SyntDagpengerApplication;
-import no.nav.dolly.synt.dagpenger.bucket.BucketUtils;
+import no.nav.dolly.synt.dagpenger.models.BucketModels;
 import no.nav.dolly.synt.dagpenger.dto.DagpengevedtakDto;
 import org.springframework.context.annotation.Profile;
 import org.springframework.stereotype.Service;
@@ -20,15 +20,15 @@ import java.util.stream.Collectors;
 class OnnxBucketService implements OnnxService {
 
     private final SyntDagpengerApplication.Config config;
-    private DagpengevedtakGenerator dagpenger;
+    private DagpengevedtakGenerator generator;
 
     @PostConstruct
-    void initialize()
+    void postConstruct()
             throws Exception {
 
-        var modelDirectory = BucketUtils.download(config.getBucket(), config.getModels(), "synt-dagpenger-models-");
-        this.dagpenger = new DagpengevedtakGenerator(modelDirectory);
-        log.info("Successfully initialized ONNX models from GCS bucket: {}", config.getBucket());
+        var modelDirectory = BucketModels.get(config.getBucket(), config.getModels(), "synt-dagpenger-models-");
+        this.generator = new DagpengevedtakGenerator(modelDirectory);
+        log.info("Successfully initialized ONNX models from bucket {}", config.getBucket());
 
     }
 
@@ -36,7 +36,7 @@ class OnnxBucketService implements OnnxService {
     public List<DagpengevedtakDto> generateVedtak(String rettighet, List<String> startDates) {
 
         var rettighetType = RettighetType.valueOf(rettighet.toUpperCase(Locale.ROOT));
-        return dagpenger
+        return generator
                 .generateVedtak(rettighetType, startDates)
                 .stream()
                 .map(VedtakMapper::fromPrediction)
