+++ synt/synt-dagpenger/src/main/java/no/nav/dolly/synt/dagpenger/onnx/OnnxLocalService.java
@@ -1,15 +1,12 @@
 package no.nav.dolly.synt.dagpenger.onnx;
 
+import jakarta.annotation.PostConstruct;
 import lombok.extern.slf4j.Slf4j;
 import no.nav.dolly.synt.dagpenger.dto.DagpengevedtakDto;
+import no.nav.dolly.synt.dagpenger.models.LocalModels;
 import org.springframework.context.annotation.Profile;
-import org.springframework.core.io.support.PathMatchingResourcePatternResolver;
 import org.springframework.stereotype.Service;
 
-import java.io.IOException;
-import java.io.UncheckedIOException;
-import java.nio.file.Files;
-import java.nio.file.Path;
 import java.util.List;
 import java.util.Locale;
 import java.util.stream.Collectors;
@@ -19,42 +16,14 @@ import java.util.stream.Collectors;
 @Slf4j
 class OnnxLocalService implements OnnxService {
 
-    private final DagpengevedtakGenerator dagpenger = new DagpengevedtakGenerator(modelDirectoryFromClasspath());
+    private DagpengevedtakGenerator generator;
 
-    private static Path modelDirectoryFromClasspath() {
+    @PostConstruct
+    void postConstruct() {
 
-        try {
-            var target = extractModelsToTemporaryDirectory();
-            log.info("Extracted ONNX models to {}", target);
-            return target;
-        } catch (IOException e) {
-            throw new UncheckedIOException("Failed to extract ONNX models from classpath", e);
-        }
-
-    }
-
-    private static Path extractModelsToTemporaryDirectory()
-            throws IOException, IllegalStateException {
-
-        var targetDir = Files.createTempDirectory("synt-dagpenger-models-");
-        targetDir.toFile().deleteOnExit();
-
-        var resolver = new PathMatchingResourcePatternResolver();
-        var resources = resolver.getResources("classpath*:models/*.onnx");
-        if (resources.length == 0) {
-            throw new IllegalStateException("No ONNX model files found at classpath:models/*.onnx");
-        }
-        for (var resource : resources) {
-            var filename = resource.getFilename();
-            if (filename == null) {
-                continue;
-            }
-            try (var in = resource.getInputStream()) {
-                Files.copy(in, targetDir.resolve(filename));
-            }
-            log.debug("Extracted model: {}", filename);
-        }
-        return targetDir;
+        var modelDirectory = LocalModels.get();
+        this.generator = new DagpengevedtakGenerator(modelDirectory);
+        log.info("Successfully initialized ONNX models from folder {}", modelDirectory.toAbsolutePath());
 
     }
 
@@ -62,7 +31,7 @@ class OnnxLocalService implements OnnxService {
     public List<DagpengevedtakDto> generateVedtak(String rettighet, List<String> startDates) {
 
         var rettighetType = RettighetType.valueOf(rettighet.toUpperCase(Locale.ROOT));
-        return dagpenger.generateVedtak(rettighetType, startDates)
+        return generator.generateVedtak(rettighetType, startDates)
                 .stream()
                 .map(VedtakMapper::fromPrediction)
                 .collect(Collectors.toList());
